operator: move types to their own package to avoid dependency loop
This allows the pkg/resources package to use the admin API internal interface.
