+++ src/go/k8s/pkg/resources/types/tls_types.go
@@ -7,7 +7,8 @@
 // the Business Source License, use of this software will be governed
 // by the Apache License, Version 2.0
 
-package resources
+// Package types contains useful types for dealing with resources
+package types
 
 import (
 	"context"
